Rename E5R.Test.Utils to E5R.Test.Commons
