+++ test/E5R.Sdk.Bit.Test/Extensions/EnumExtensionsTests.cs
@@ -7,10 +7,10 @@
 using System;
 using Xunit;
 
-namespace E5R.Sdk.Bit.Test
+namespace E5R.Sdk.Bit.Test.Extensions
 {
-    using Utils;
-    using _ = E5R.Test.Utils.TraitConstants;
+    using Bit.Utils;
+    using _ = E5R.Test.Commons.TraitConstants;
 
     [Trait(_.MODULE, "E5R.Sdk.Bit")]
     [Trait(_.COMPONENT, nameof(EnumExtensions))]
