+++ test/E5R.Sdk.Bit.Test/Utils/BitUriTests.cs
@@ -4,10 +4,10 @@
 using System;
 using Xunit;
 
-namespace E5R.Sdk.Bit.Test
+namespace E5R.Sdk.Bit.Test.Utils
 {
-    using Utils;
-    using _ = E5R.Test.Utils.TraitConstants;
+    using Bit.Utils;
+    using _ = E5R.Test.Commons.TraitConstants;
 
     [Trait(_.MODULE, "E5R.Sdk.Bit")]
     [Trait(_.COMPONENT, nameof(BitUri))]
