+++ test/E5R.Test.Commons/E5R.Test.Commons.csproj
@@ -3,8 +3,8 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <IsPackable>false</IsPackable>
-    <AssemblyName>E5R.Test.Utils</AssemblyName>    
-    <RootNamespace>E5R.Test.Utils</RootNamespace>
+    <AssemblyName>E5R.Test.Commons</AssemblyName>    
+    <RootNamespace>E5R.Test.Commons</RootNamespace>
   </PropertyGroup>
 
 </Project>
