+++ test/E5R.Test.Commons/TraitConstants.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) E5R Development Team. All rights reserved.
 // Licensed under the Apache License, Version 2.0. More license information in LICENSE.txt.
 
-namespace E5R.Test.Utils
+namespace E5R.Test.Commons
 {
     public class TraitConstants
     {
