+++ test/E5R.Tools.Bit.Engine.Test/E5R.Tools.Bit.Engine.Test.csproj
@@ -3,19 +3,20 @@
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
     <IsPackable>false</IsPackable>
-    <AssemblyName>E5R.Tools.Bit.Engine.Test</AssemblyName>    
+    <AssemblyName>E5R.Tools.Bit.Engine.Test</AssemblyName>
     <RootNamespace>E5R.Tools.Bit.Engine.Test</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0-preview-20170628-02" />
+    <PackageReference Include="Moq" Version="4.7.142" />
     <PackageReference Include="xunit" Version="2.2.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\E5R.Tools.Bit.Engine\E5R.Tools.Bit.Engine.csproj" />
-    <ProjectReference Include="..\E5R.Test.Utils\E5R.Test.Utils.csproj" />
+    <ProjectReference Include="..\E5R.Test.Commons\E5R.Test.Commons.csproj" />
   </ItemGroup>
-
+  
 </Project>
